Resolvendo conflitos da classe Professor.java e RelatorioFinanceiro.java

+++ src/main/java/com/fitlife/Professor/Professor.java
@@ -2,9 +2,9 @@ package com.fitlife.Professor;
 
 public class Professor {
 
-    private Long id; private String nome; private String registro; private String especializacao;
+    private int id; private String nome; private String registro; private String especializacao;
 
-    public Professor(Long id, String nome, String registro, String especializacao) {
+    public Professor(int id, String nome, String registro, String especializacao) {
         this.id = id;
         this.nome = nome;
         this.registro = registro;
@@ -14,7 +14,7 @@ public class Professor {
     public Professor(String csvLine) {
         String[] dados = csvLine.split(";");
         if (dados.length == 4) {
-            this.id = Long.parseLong(dados[0].trim());
+            this.id = Integer.parseInt(dados[0].trim());
             this.nome = dados[1].trim();
             this.registro = dados[2].trim();
             this.especializacao = dados[3].trim();
@@ -23,11 +23,11 @@ public class Professor {
         }
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
